Added Access Point script for the RPi (#18)
* Added Access Point script for the RPi

The script prompts if it should be enabled or disabled. The default IP of the Pi is 192.168.0.1 and it can only run in Access Point mode or in the default mode (in order to connect to the internet). It has to be run with bash and using sudo or as root.

* No longer requires sudo on the script

Now sudo is just inside the script, it is safer and more secure this way. Really the other way just isn't a good idea.

* Added case-insensitive yes/no

* Update wifi_enable.sh

Shortened the SSID just in case there is a long hostname.
Added flags/options.
Change the first `ifdown` to `--force` because otherwise it sometimes has issues. I'm not sure if it could cause issues though.